requirements.txt
fastapi==0.87.0
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pytest==7.2.0
python-dotenv==0.21.0
SQLAlchemy==1.4.43
